added a reading route + redirect user to reading when in /
