+++ src/app/page.tsx
@@ -1,101 +1,17 @@
 "use client";
 
-import Sidebar from "@/components/features/Sidebar/Sidebar";
-import { readingTopics } from "@/data/topics";
-import BookSVG from "@/components/features/Questions/icons/BookSVG";
-import { Topic } from "@/types/topic";
-import { useState, useEffect, useRef } from "react";
-import ReadingQuestion from "@/components/features/Questions/Question-UI/ReadingQuestion";
-import Header from "@/components/features/Questions/Header";
-import { useAnswerCounterStore } from "@/store/score";
-import ScoreModal from "@/components/features/Questions/Modals/ScoreModal";
-import StreakModal from "@/components/features/Questions/Modals/StreakModal";
-import { Answers } from "@/types/answer";
-import { useScoreModalStore, useStreakAnnouncerModalStore, useStreakCounterModalStore } from "@/store/modals";
-import StreakAnnouncer from "@/components/features/Questions/Modals/StreakAnnouncer";
-import useQuestionHandler from "@/hooks/questions";
-import Spinner from "@/components/common/Spinner";
-import GetStarted from "@/components/features/Questions/GetStarted";
-import Result from "@/components/features/Questions/Results";
-import { useQuestionStore } from "@/store/questions";
-import QuestionWrappers from "@/components/wrappers/question/Question";
-import MainWrappers from "@/components/wrappers/question/Main";
+import { useEffect } from "react";
+import { useRouter } from "next/router";
 
 const Reading = () => {
-  const { fetchRandomQuestion, handleAnswerSubmit, handleCheckThreeStreak } = useQuestionHandler();
-  const [selectedTopic, setSelectedTopic] = useState<Topic | null>(null);
-  const randomQuestion = useQuestionStore((state) => state.randomQuestion);
-  const correctCount = useAnswerCounterStore((state) => state.count);
-  const answerCorrectRef: Record<Answers, number> = { A: 0, B: 1, C: 2, D: 3 };
-
-  const isScoreModalOpen = useScoreModalStore((state) => state.isOpen);
-  const isAnnouncerModalOpen = useStreakAnnouncerModalStore((state) => state.isOpen);
-  const isStreakModalOpen = useStreakCounterModalStore((state) => state.isOpen);
-
-  const answerComponent = useRef<HTMLDivElement | null>(null);
+  const router = useRouter();
 
   useEffect(() => {
-    handleCheckThreeStreak();
-  }, [correctCount, handleCheckThreeStreak]);
+    // Redirect to the /reading route
+    router.push("/reading");
+  }, [router]);
 
-  const handleTopicClick = (topic: Topic) => {
-    setSelectedTopic(topic);
-    fetchRandomQuestion("Reading", topic);
-  };
-
-  if (isScoreModalOpen || isStreakModalOpen) {
-    return (
-      <>
-        <ScoreModal />
-        <StreakModal />
-      </>
-    );
-  }
-
-  return (
-    <MainWrappers>
-      <Sidebar
-        title="Reading"
-        svg={<BookSVG />}
-        topics={readingTopics}
-        selectedTopic={selectedTopic!}
-        handleTopicClick={handleTopicClick}
-      />
-
-      {/* Main Content */}
-      <QuestionWrappers>
-        {selectedTopic ? (
-          <div className="w-full mx-auto">
-            <Header
-              name={selectedTopic.name}
-              question={randomQuestion?.question}
-            />
-            {randomQuestion ? (
-              <ReadingQuestion
-                onAnswerSubmit={() =>
-                  handleAnswerSubmit(
-                    "Reading",
-                    randomQuestion.correctAnswer,
-                    answerCorrectRef
-                  )
-                }
-              />
-            ) : (
-              <Spinner />
-            )}
-            <Result
-              answerComponent={answerComponent}
-              explanation={randomQuestion?.explanation || ""}
-            />
-          </div>
-        ) : (
-          <GetStarted />
-        )}
-      </QuestionWrappers>
-
-      {isAnnouncerModalOpen && <StreakAnnouncer />}
-    </MainWrappers>
-  );
+  return null; // Render nothing as the user is being redirected
 };
 
 export default Reading;
